pick_and_play_animation is headerless
