+++ src/Samples/src/samples/animations/_animations_samples_index.cpp
@@ -2,7 +2,6 @@
 
 #include <babylon/samples/animations/easing/easing_functions_scene.h>
 #include <babylon/samples/animations/easing/shape_easings_scene.h>
-#include <babylon/samples/animations/pick_and_play_animation.h>
 #include <babylon/samples/animations/pump_jack_scene.h>
 #include <babylon/samples/animations/tube_animation_scene.h>
 
@@ -14,10 +13,6 @@ _AnimationsSamplesIndex::_AnimationsSamplesIndex()
 {
   // -- Animations examples --
 
-  // Pick And Play Animation Scene
-  _samples["PickAndPlayAnimationScene"] = [](ICanvas* iCanvas) {
-    return std::make_unique<PickAndPlayAnimationScene>(iCanvas);
-  };
   // Pump Jack Scene
   _samples["PumpJackScene"]
     = [](ICanvas* iCanvas) { return std::make_unique<PumpJackScene>(iCanvas); };
